requirements.txt
pandas>=1.5.0
numpy>=1.21.0
openai>=1.0.0
requests>=2.28.0
regex>=2022.7.9
chardet>=5.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
